{"schema":"decodex/commit/1","summary":"Split MCP autonomy resource modules","authority":"XY-1115"}
