+++ apps/decodex/src/mcp/autonomy_resources.rs
@@ -1,338 +1,14 @@
-use serde_json::{self, Value};
+mod resources;
+mod summaries;
 
-use crate::{
-	autonomy_objective::AutonomyObjectiveContract,
-	autonomy_proposal::AutonomyProposal,
-	autonomy_signal::{AutonomySignal, AutonomySignalPrivacy},
-	state::{AutonomyProposalRecord, AutonomySignalRecord, StateStore},
+pub(super) use self::{
+	resources::{
+		mcp_autonomy_current_objective_resource, mcp_autonomy_evidence_resource,
+		mcp_autonomy_objective_version_resource, mcp_autonomy_project_resource,
+		mcp_autonomy_proposal_resource, mcp_autonomy_proposals_resource,
+		mcp_autonomy_signal_resource, mcp_autonomy_signals_resource,
+	},
+	summaries::{
+		mcp_autonomy_objective_summary, mcp_autonomy_proposal_summary, mcp_autonomy_signal_summary,
+	},
 };
-
-use super::{DEFAULT_MCP_STATUS_LIMIT, McpError};
-
-pub(super) fn mcp_autonomy_project_resource(
-	state_store: &StateStore,
-	project_id: &str,
-) -> crate::prelude::Result<Value, McpError> {
-	let objectives = state_store
-		.recent_autonomy_objectives_for_project(project_id, DEFAULT_MCP_STATUS_LIMIT)
-		.map_err(McpError::internal)?;
-	let signals = state_store
-		.recent_autonomy_signals_for_project(project_id, DEFAULT_MCP_STATUS_LIMIT)
-		.map_err(McpError::internal)?;
-	let proposals = state_store
-		.recent_autonomy_proposals_for_project(project_id, DEFAULT_MCP_STATUS_LIMIT)
-		.map_err(McpError::internal)?;
-
-	Ok(serde_json::json!({
-		"schema": "decodex.mcp.autonomy_summary/1",
-		"project_id": project_id,
-		"read_only": true,
-		"authority_boundary": mcp_autonomy_authority_boundary(),
-		"objectives": objectives
-			.iter()
-			.map(|record| mcp_autonomy_objective_summary(record.objective(), Some(record.updated_at())))
-			.collect::<Vec<_>>(),
-		"signals": signals
-			.iter()
-			.map(|record| mcp_autonomy_signal_summary(record.signal(), Some(record.updated_at())))
-			.collect::<Vec<_>>(),
-		"proposals": proposals
-			.iter()
-			.map(|record| mcp_autonomy_proposal_summary(record.proposal(), Some(record.updated_at())))
-			.collect::<Vec<_>>(),
-		"evidence": mcp_autonomy_evidence_summary(&signals, &proposals)
-	}))
-}
-
-pub(super) fn mcp_autonomy_current_objective_resource(
-	state_store: &StateStore,
-	project_id: &str,
-	objective_id: &str,
-) -> crate::prelude::Result<Value, McpError> {
-	let Some(record) = state_store
-		.current_accepted_autonomy_objective(project_id, objective_id)
-		.map_err(McpError::internal)?
-	else {
-		return Err(McpError::resource_not_found());
-	};
-
-	Ok(serde_json::json!({
-		"schema": "decodex.mcp.autonomy_objective_resource/1",
-		"project_id": project_id,
-		"read_only": true,
-		"authority_boundary": mcp_autonomy_authority_boundary(),
-		"objective": mcp_autonomy_objective_summary(record.objective(), Some(record.updated_at()))
-	}))
-}
-
-pub(super) fn mcp_autonomy_objective_version_resource(
-	state_store: &StateStore,
-	project_id: &str,
-	objective_id: &str,
-	version: &str,
-) -> crate::prelude::Result<Value, McpError> {
-	let version = version.parse::<u64>().map_err(|_| McpError::resource_not_found())?;
-	let Some(record) = state_store
-		.autonomy_objective(project_id, objective_id, version)
-		.map_err(McpError::internal)?
-	else {
-		return Err(McpError::resource_not_found());
-	};
-
-	Ok(serde_json::json!({
-		"schema": "decodex.mcp.autonomy_objective_resource/1",
-		"project_id": project_id,
-		"read_only": true,
-		"authority_boundary": mcp_autonomy_authority_boundary(),
-		"objective": mcp_autonomy_objective_summary(record.objective(), Some(record.updated_at()))
-	}))
-}
-
-pub(super) fn mcp_autonomy_signals_resource(
-	state_store: &StateStore,
-	project_id: &str,
-) -> crate::prelude::Result<Value, McpError> {
-	let signals = state_store
-		.recent_autonomy_signals_for_project(project_id, DEFAULT_MCP_STATUS_LIMIT)
-		.map_err(McpError::internal)?;
-
-	Ok(serde_json::json!({
-		"schema": "decodex.mcp.autonomy_signals_resource/1",
-		"project_id": project_id,
-		"read_only": true,
-		"signals": signals
-			.iter()
-			.map(|record| mcp_autonomy_signal_summary(record.signal(), Some(record.updated_at())))
-			.collect::<Vec<_>>()
-	}))
-}
-
-pub(super) fn mcp_autonomy_signal_resource(
-	state_store: &StateStore,
-	project_id: &str,
-	signal_id: &str,
-) -> crate::prelude::Result<Value, McpError> {
-	let Some(record) =
-		state_store.autonomy_signal(project_id, signal_id).map_err(McpError::internal)?
-	else {
-		return Err(McpError::resource_not_found());
-	};
-
-	Ok(serde_json::json!({
-		"schema": "decodex.mcp.autonomy_signal_resource/1",
-		"project_id": project_id,
-		"read_only": true,
-		"signal": mcp_autonomy_signal_summary(record.signal(), Some(record.updated_at()))
-	}))
-}
-
-pub(super) fn mcp_autonomy_proposals_resource(
-	state_store: &StateStore,
-	project_id: &str,
-) -> crate::prelude::Result<Value, McpError> {
-	let proposals = state_store
-		.recent_autonomy_proposals_for_project(project_id, DEFAULT_MCP_STATUS_LIMIT)
-		.map_err(McpError::internal)?;
-
-	Ok(serde_json::json!({
-		"schema": "decodex.mcp.autonomy_proposals_resource/1",
-		"project_id": project_id,
-		"read_only": true,
-		"proposals": proposals
-			.iter()
-			.map(|record| mcp_autonomy_proposal_summary(record.proposal(), Some(record.updated_at())))
-			.collect::<Vec<_>>()
-	}))
-}
-
-pub(super) fn mcp_autonomy_proposal_resource(
-	state_store: &StateStore,
-	project_id: &str,
-	proposal_id: &str,
-) -> crate::prelude::Result<Value, McpError> {
-	let Some(record) =
-		state_store.autonomy_proposal(project_id, proposal_id).map_err(McpError::internal)?
-	else {
-		return Err(McpError::resource_not_found());
-	};
-
-	Ok(serde_json::json!({
-		"schema": "decodex.mcp.autonomy_proposal_resource/1",
-		"project_id": project_id,
-		"read_only": true,
-		"proposal": mcp_autonomy_proposal_summary(record.proposal(), Some(record.updated_at()))
-	}))
-}
-
-pub(super) fn mcp_autonomy_evidence_resource(
-	state_store: &StateStore,
-	project_id: &str,
-) -> crate::prelude::Result<Value, McpError> {
-	let signals = state_store
-		.recent_autonomy_signals_for_project(project_id, DEFAULT_MCP_STATUS_LIMIT)
-		.map_err(McpError::internal)?;
-	let proposals = state_store
-		.recent_autonomy_proposals_for_project(project_id, DEFAULT_MCP_STATUS_LIMIT)
-		.map_err(McpError::internal)?;
-
-	Ok(serde_json::json!({
-		"schema": "decodex.mcp.autonomy_evidence_resource/1",
-		"project_id": project_id,
-		"read_only": true,
-		"evidence": mcp_autonomy_evidence_summary(&signals, &proposals)
-	}))
-}
-
-pub(super) fn mcp_autonomy_objective_summary(
-	objective: &AutonomyObjectiveContract,
-	updated_at: Option<&str>,
-) -> Value {
-	serde_json::json!({
-		"objective_id": objective.id(),
-		"objective_version": objective.version(),
-		"state": objective.state().as_str(),
-		"summary": objective.summary(),
-		"goals": objective.goals(),
-		"non_goals": objective.non_goals(),
-		"metrics": objective.metrics(),
-		"allowed_surfaces": objective.allowed_surfaces(),
-		"allowed_signal_kinds": objective.allowed_signal_kinds(),
-		"validation_gates": objective.validation_gates(),
-		"review_policy": objective.review_policy(),
-		"acceptance_present": objective.acceptance().is_some(),
-		"updated_at": updated_at
-	})
-}
-
-pub(super) fn mcp_autonomy_signal_summary(
-	signal: &AutonomySignal,
-	updated_at: Option<&str>,
-) -> Value {
-	let (source_refs, primary_source_refs, source_ref_count, primary_source_ref_count) =
-		mcp_autonomy_signal_ref_summary(signal);
-
-	serde_json::json!({
-		"signal_id": signal.id(),
-		"objective_id": signal.objective_id(),
-		"objective_version": signal.objective_version(),
-		"kind": signal.kind().as_str(),
-		"source_type": signal.source_type().as_str(),
-		"source_refs": source_refs,
-		"source_ref_count": source_ref_count,
-		"primary_source_refs": primary_source_refs,
-		"primary_source_ref_count": primary_source_ref_count,
-		"freshness": signal.freshness().as_str(),
-		"summary": signal.summary(),
-		"evidence_class": signal.evidence_class().as_str(),
-		"confidence": signal.confidence().as_str(),
-		"redaction_level": signal.privacy().as_str(),
-		"gaps": signal.gaps(),
-		"contradictions": signal.contradictions(),
-		"review_evidence_present": signal.review_evidence().is_some(),
-		"updated_at": updated_at
-	})
-}
-
-fn mcp_autonomy_signal_ref_summary(signal: &AutonomySignal) -> (Value, Value, usize, usize) {
-	let source_ref_count = signal.source_refs().len();
-	let primary_source_ref_count = signal.primary_source_refs().len();
-
-	if signal.privacy() == AutonomySignalPrivacy::LocalPrivate {
-		return (
-			serde_json::json!([]),
-			serde_json::json!([]),
-			source_ref_count,
-			primary_source_ref_count,
-		);
-	}
-
-	(
-		serde_json::json!(signal.source_refs()),
-		serde_json::json!(signal.primary_source_refs()),
-		source_ref_count,
-		primary_source_ref_count,
-	)
-}
-
-pub(super) fn mcp_autonomy_proposal_summary(
-	proposal: &AutonomyProposal,
-	updated_at: Option<&str>,
-) -> Value {
-	serde_json::json!({
-		"proposal_id": proposal.id(),
-		"objective_id": proposal.objective_id(),
-		"objective_version": proposal.objective_version(),
-		"state": proposal.state().as_str(),
-		"summary": proposal.summary(),
-		"source_family": proposal.source_family(),
-		"intended_surface": proposal.intended_surface(),
-		"affected_identifiers": proposal.affected_identifiers(),
-		"source_signal_ids": proposal.source_signal_ids(),
-		"allowed_surfaces": proposal.allowed_surfaces(),
-		"validation_gates": proposal.validation_gates(),
-		"refusal_reasons": proposal
-			.refusal_reasons()
-			.iter()
-			.map(|refusal| refusal.reason().as_str())
-			.collect::<Vec<_>>(),
-		"refusals": proposal
-			.refusal_reasons()
-			.iter()
-			.map(|refusal| {
-				serde_json::json!({
-					"reason": refusal.reason().as_str(),
-					"detail": refusal.detail(),
-					"evidence_refs": refusal.evidence_refs()
-				})
-			})
-			.collect::<Vec<_>>(),
-		"gaps": proposal.gaps(),
-		"contradictions": proposal.contradictions(),
-		"challenge_evidence_count": proposal.challenge_evidence().len(),
-		"updated_at": updated_at
-	})
-}
-
-fn mcp_autonomy_evidence_summary(
-	signals: &[AutonomySignalRecord],
-	proposals: &[AutonomyProposalRecord],
-) -> Value {
-	serde_json::json!({
-		"signal_count": signals.len(),
-		"proposal_count": proposals.len(),
-		"signal_refs": signals
-			.iter()
-			.map(|record| {
-				serde_json::json!({
-					"signal_id": record.signal_id(),
-					"kind": record.kind().as_str(),
-					"freshness": record.freshness().as_str(),
-					"evidence_class": record.evidence_class().as_str(),
-					"confidence": record.confidence().as_str(),
-					"redaction_level": record.privacy().as_str()
-				})
-			})
-			.collect::<Vec<_>>(),
-		"proposal_refs": proposals
-			.iter()
-			.map(|record| {
-				serde_json::json!({
-					"proposal_id": record.proposal_id(),
-					"state": record.state().as_str(),
-					"objective_id": record.objective_id(),
-					"objective_version": record.objective_version()
-				})
-			})
-			.collect::<Vec<_>>(),
-		"authority_effect": "evidence_summary_only_no_execution_authority"
-	})
-}
-
-fn mcp_autonomy_authority_boundary() -> Value {
-	serde_json::json!({
-		"mcp_authentication": "access_boundary_only",
-		"capability_profile": "tool_visibility_boundary_only",
-		"acceptance_authority": "explicit_human_or_trusted_accepted_project_policy_required",
-		"execution_authority": "Decision Contract promotion and Program Intake remain separate"
-	})
-}
